chatbot-ai/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
databases==0.8.0
distro==1.8.0
exceptiongroup==1.2.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
load-dotenv==0.1.0
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
openai==1.5.0
passlib==1.7.4
pgvector==0.2.4
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
